+++ libs/community/langchain_community/memory/zep_memory.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any, Dict, Optional
 
-from langchain.memory import ConversationBufferMemory
+from langchain_core.memory import ConversationBufferMemory
 
 from langchain_community.chat_message_histories import ZepChatMessageHistory
